Moved over all declarations for chat and player into the components. I also figured out why there wasn't another page being created when testing the chat function
